+++ egs2/wsj0_Nmix/enh_tse1_max/local/data.sh
@@ -29,11 +29,11 @@ wsj_scp_output_dir=$PWD/data
 
 other_text=data/local/other_text/text
 nlsyms=data/nlsyms.txt
-min_or_max=max
+min_or_max=min
 sample_rate=8k
 
-stage=4
-stop_stage=4
+stage=3
+stop_stage=100
 
 . utils/parse_options.sh
 
@@ -61,34 +61,41 @@ if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     echo "Downloading WSJ0_mixture scripts."
     mkdir -p ${wsj_mix_scripts}
 
-    git clone https://github.com/mpariente/pywsj0-mix.git ${wsj_mix_scripts}
-    cp local/generate_wsj_scp.py ${wsj_mix_scripts}
+    if [ ! -e ${wsj_mix_scripts} ]; then
+        git clone https://github.com/mpariente/pywsj0-mix.git ${wsj_mix_scripts}
+        cp local/create_wsj_scp.py ${wsj_mix_scripts}
 
-    echo "WSJ0 wav file."
-    local/convert2wav.sh ${WSJ0} ${wsj_full_wav} || exit 1;
+        # remove original script and copy multi-thred version to speed-up data preparation
+        rm ${wsj_mix_scripts}/generate_wsjmix.py
+        cp local/generate_wsjmix.py ${wsj_mix_scripts}
+    else
+        echo "${wsj_mix_scripts} already exists. Skip downloading the simulation script."
+    fi
+
+    if [ ! -e ${wsj_full_wav} ]; then
+        echo "WSJ0 wav file."
+        local/convert2wav.sh ${WSJ0} ${wsj_full_wav} || exit 1;
+    else
+        echo "${wsj_full_wav} already exists. Skip the process to convert wsj0 to wav."
+    fi
 fi
 
 if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
     log "stage 1: Data simulation"
-    ### This part is for WSJ0 mix
-    for nsrc in 2 3 4 5;
-    do
+    for nsrc in 2 3 4 5; do
+        # create mixtures
         local/wsj0_create_mixture.sh --min-or-max ${min_or_max} --sample-rate ${sample_rate} \
             ${wsj_mix_scripts} ${WSJ0} ${wsj_full_wav} \
             ${wsj_mix_wav} ${nsrc} || exit 1;
+        # create scp files (unsorted yet)
         local/wsj0_create_scp.sh --min-or-max ${min_or_max} --sample-rate ${sample_rate} \
             ${wsj_mix_scripts} ${WSJ0} ${wsj_full_wav} \
             ${wsj_mix_wav} ${nsrc} ${wsj_scp_output_dir} || exit 1;
     done
+    # sort scp files and create utt2category files
+    # also prepare transcriptions for max version for ASR evaluation
     local/wsj0mix_data_prep.sh --min-or-max ${min_or_max} --sample-rate ${sample_rate} \
         ${wsj_mix_wav}/wav${sample_rate}/${min_or_max} ${wsj_mix_scripts} ${wsj_full_wav} || exit 1;
-
-    for task in tr cv;
-    do
-        python local/prepare_utt2category.py \
-        ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate}/wav.scp \
-        --output_dir ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate} || exit 1;
-    done
 fi
 
 if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
@@ -100,55 +107,59 @@ if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
     log "local/wsj_format_data.sh"
     local/wsj_format_data.sh
     log "mkdir -p data/wsj"
-    mkdir -p data/wsj
-    log "mv data/{dev_dt_*,local,test_dev*,test_eval*,train_si284} data/wsj"
-    mv data/{dev_dt_*,local,test_dev*,test_eval*,train_si284} data/wsj
 
-    log "Prepare text from lng_modl dir: ${WSJ1}/13-32.1/wsj1/doc/lng_modl/lm_train/np_data/{87,88,89}/*.z -> ${other_text}"
-    mkdir -p "$(dirname ${other_text})"
+    if [ ! -e "data/wsj" ]; then
+        mkdir -p data/wsj
+        log "mv data/{dev_dt_*,local,test_dev*,test_eval*,train_si284} data/wsj"
+        mv data/{dev_dt_*,local,test_dev*,test_eval*,train_si284} data/wsj
+    fi
 
-    # NOTE(kamo): Give utterance id to each texts.
-    zcat ${WSJ1}/13-32.1/wsj1/doc/lng_modl/lm_train/np_data/{87,88,89}/*.z | \
-        grep -v "<" | tr "[:lower:]" "[:upper:]" | \
-        awk '{ printf("wsj1_lng_%07d %s\n",NR,$0) } ' > ${other_text}
+    if [ ! -e ${other_text} ]; then
+        log "Prepare text from lng_modl dir: ${WSJ1}/13-32.1/wsj1/doc/lng_modl/lm_train/np_data/{87,88,89}/*.z -> ${other_text}"
+        mkdir -p "$(dirname ${other_text})"
 
+        # NOTE(kamo): Give utterance id to each texts.
+        zcat ${WSJ1}/13-32.1/wsj1/doc/lng_modl/lm_train/np_data/{87,88,89}/*.z | \
+            grep -v "<" | tr "[:lower:]" "[:upper:]" | \
+            awk '{ printf("wsj1_lng_%07d %s\n",NR,$0) } ' > ${other_text}
+    fi
 
-
-    log "Create non linguistic symbols: ${nlsyms}"
-    cut -f 2- data/wsj/train_si284/text | tr " " "\n" | sort | uniq | grep "<" > ${nlsyms}
-    cat ${nlsyms}
+    if [ ! -e ${nlsyms} ]; then
+        log "Create non linguistic symbols: ${nlsyms}"
+        cut -f 2- data/wsj/train_si284/text | tr " " "\n" | sort | uniq | grep "<" > ${nlsyms}
+        cat ${nlsyms}
+    fi
 fi
 
 if [ ${stage} -le 3 ] && [ ${stop_stage} -ge 3 ]; then
     log "stage 3: Prepare WSJMix enroll utterances"
 
-    for task in tr cv;
+    for task in tr cv tt;
     do
-        python local/prepare_spk2enroll_wsj.py \
-            ${wsj_mix_wav}/5speakers/wav${sample_rate}/${min_or_max}/${task} \
-            --output_dir ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate} || exit 1;
-    done
+        # copy scp files from local while fixing some infomation
+        # e.g., min->max, 8k->16k and make them absolute paths
+        python local/copy_enroll_files.py \
+            --source_folder ./local/enroll_scp/${task}_min_8k \
+            --target_folder ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate} \
+            --source_words ./ /min/ /wav8k/ \
+            --target_words ${PWD}/ /${min_or_max}/ /wav${sample_rate}/
 
-    for task in tr cv;
-    do
+        # for training set, we prepare scp files here
         if [ "$task" = "tr" ]; then
-            train=true
-        else
-            train=false
+            python local/prepare_wsj_enroll.py \
+                ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate}/wav.scp \
+                ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate}/spk2enroll.json \
+                --num_spk 5 --train true \
+                --output_dir ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate} || exit 1;
         fi
 
-        python local/prepare_wsj_enroll.py \
-            ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate}/wav.scp \
-            ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate}/spk2enroll.json \
-            --num_spk 5 --train ${train} --seed 1 \
-            --output_dir ${wsj_scp_output_dir}/${task}_${min_or_max}_${sample_rate} || exit 1;
     done
 fi
 
 
-if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
-    log "stage 4: Prepare 1-mix data"
-    ${PWD}/local/prepare_1speaker_audio_and_scp.sh \
-        --min_or_max ${min_or_max} --sample_rate ${sample_rate}
-fi
+# if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
+#     log "stage 4: Prepare 1-mix data"
+#     ${PWD}/local/prepare_1speaker_audio_and_scp.sh \
+#         --min_or_max ${min_or_max} --sample_rate ${sample_rate}
+# fi
